move doc into doc folder
