+++ package.json
@@ -27,6 +27,19 @@
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "7.3.3",
     "@babel/register": "7.4.4",
+    "@storybook/addon-a11y": "^4.1.3",
+    "@storybook/addon-actions": "^4.1.3",
+    "@storybook/addon-console": "^1.1.0",
+    "@storybook/addon-info": "^4.1.4",
+    "@storybook/addon-knobs": "^4.1.3",
+    "@storybook/addon-links": "^4.1.3",
+    "@storybook/addon-notes": "^4.1.3",
+    "@storybook/addon-options": "^4.1.4",
+    "@storybook/addon-storysource": "^4.1.3",
+    "@storybook/addon-viewport": "^4.1.3",
+    "@storybook/addons": "^4.1.3",
+    "@storybook/cli": "^4.1.3",
+    "@storybook/react": "^4.1.3",
     "@typescript-eslint/eslint-plugin": "^1.4.2",
     "@typescript-eslint/parser": "^1.4.2",
     "ajv": "6.10.0",
@@ -76,6 +89,7 @@
     "signale": "1.4.0",
     "slash2": "2.0.0",
     "staged-git-files": "^1.2.0",
+    "storybook-addon-source": "latest",
     "temp-dir": "2.0.0",
     "through2": "3.0.1",
     "typescript": "3.5.1",
