+++ src/doc/doc.test.ts
@@ -1,13 +1,13 @@
 import { join } from 'path';
 import { existsSync, renameSync } from 'fs';
-import { devOrBuild } from './doc';
+import { devOrBuild } from './';
 
 xdescribe('father doc build', () => {
   process.env.COMPRESS = 'none';
   require('test-build-result')({
-    root: join(__dirname, './fixtures/doc'),
+    root: join(__dirname, '../fixtures/doc'),
     build({ cwd }) {
-      return devOrBuild({ cwd, cmd: 'build', params: [], docConfig: {} }).then(() => {
+      return devOrBuild({ cwd, cmd: 'build', params: [], docConfig: {} } as any).then(() => {
         const absDirPath = join(cwd, '.docz/dist');
         if (existsSync(absDirPath)) {
           renameSync(absDirPath, join(cwd, 'dist'));
